Basic weapons
Can easily create new weapons, but this lets the player have a handful of weapons and switch between them and shoot. Still need to do a little more work and there is some unity editor stuff you'll need to do like add a weapon handler gameobject with the script. I'll refine it more soon so it's easier to incorporate, but I want to get the files up here.